requirements.txt
absl-py==0.8.1
astor==0.8.0
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
decorator==4.4.1
gast==0.3.2
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.8
imagecodecs-lite==2019.12.3
imageio==2.6.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
llvmlite==0.30.0
Markdown==3.1.1
matplotlib==3.1.2
mock==3.0.5
networkx==2.4
numba==0.46.0
numpy==1.16.4
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
pandas==0.25.3
Pillow==6.2.1
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scipy==1.3.3
Shapely==1.6.4.post2
six==1.13.0
tensorboard==1.13.1
tensorflow-gpu==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
tifffile==2019.7.26.2
tqdm==4.40.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
